[jOOQ/jOOQ#13509] Apply fix also for SQL Server

+++ jOOQ/src/main/java/org/jooq/impl/SelectQueryImpl.java
@@ -355,6 +355,8 @@ final class SelectQueryImpl<R extends Record> extends AbstractResultQuery<R> imp
 
 
 
+
+
     final WithImpl                                       with;
     private final SelectFieldList<SelectFieldOrAsterisk> select;
     private Table<?>                                     intoTable;
@@ -3081,9 +3083,11 @@ final class SelectQueryImpl<R extends Record> extends AbstractResultQuery<R> imp
             if (actualLimit.isApplicable()) {
                 ctx.visit(actualLimit);
             }
+
+            // [#13509] Force a LIMIT clause to prevent optimisation of "unnecessary" ORDER BY
             else if (!actualOrderBy.isEmpty() && TRUE.equals(ctx.data(DATA_FORCE_LIMIT_WITH_ORDER_BY))) {
                 Limit l = new Limit();
-                l.setNumberOfRows(Integer.MAX_VALUE);
+                l.setNumberOfRows(Long.MAX_VALUE);
                 ctx.visit(l);
             }
         }
@@ -3103,7 +3107,6 @@ final class SelectQueryImpl<R extends Record> extends AbstractResultQuery<R> imp
 
         ;
     }
-
